Bigtable: Updated warpper api namespace and test cases
- Added google-cloud-core gem for service client creation
- Updated testcases for admin table and instance wrapper apis
- Updated docs
- Removed unused error classes
