+++ google-cloud-bigtable/google-cloud-bigtable.gemspec
@@ -22,6 +22,7 @@ Gem::Specification.new do |gem|
 
   gem.add_dependency "google-gax", "~> 1.0"
   gem.add_dependency "grpc-google-iam-v1", "~> 0.6.9"
+  gem.add_dependency "google-cloud-core", "~> 1.1"
 
   gem.add_development_dependency "minitest", "~> 5.10"
   gem.add_development_dependency "rubocop", "~> 0.50.0"
